Make SPI.h work on non-AVR unittests
The SPI library contains some code to read the SPCR register to
determine the byte order of 16-bit transfers, presumably because there
is no official Arduino API to set it, so sketches had to rely on setting
this register directly. By reading it, the SPI unittest implementation
can detect how to emulate multibyte transfers.

However, this register is only defined by avr/io.h when the unittest
emulates an AVR platform, so this code would fail to compile on other
platforms.

This adds a preprocessor guard around this code, defaulting to the
lsb-first, which is also the hardware and Arduino default.

This fixes #140.