Complete and fix defines for various boards
This adds relevant defines that identify the architecture and board
currently compiled for. Most of these are usually set by the platform's
platform.txt and boards.txt, except for the __AVR* defines that are set
by avr-gcc internally.

This only adds extra defines, except for the Arduino Due, which
previously incorrectly identified as an ATmega328p.

This seems to fix part of #89.