Remove some functions from pgmspace.h that break on cygwin
Somehow, these functions cannot be found when compiling the test build
under cygwin, even though it seems they should be available. To at least
allow the rest of the fixes in this series to be merged, disable these
functions for now. Since these are mostly uncommon functions that are
unlikely to be used in actual Arduino code, this should probably not
cause much problems for now.